bot/src/requirements.txt
aiohttp==3.7.2
alembic==1.4.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.1
idna==2.10
Mako==1.1.3
MarkupSafe==1.1.1
multidict==5.0.0
Pillow==8.0.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
pyxivapi==0.2.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
